+++ src/main/java/com/seletoai/core/ports/in/auth/CreateRefreshTokenUseCasePort.java
@@ -4,5 +4,5 @@ import com.seletoai.core.domain.auth.RefreshToken;
 
 public interface CreateRefreshTokenUseCasePort {
 
-  RefreshToken execute(Integer userId);
+  RefreshToken execute(Long userId);
 }
